Release nxtvepg-2.7.1 (2004-Jul-24)
Changes between 2.7.0 and 2.7.1 (as per file CHANGES):
- bugfix in "local time" option for XML database export (times were still
  in GMT only, regardless if the option was enabled or not)
- added support for new XMLTV DTD 0.6 (note command line option -dump xml
  uses the format which was used last in Control->"Export as XMLTV")
- fixed -daemonstop option for Linux and v4l2 devices (2.6 kernels): did
  sometimes not work after the user interface was used.
- added support for new UNIX TV app: Zapping, i.e. reading channel table
  from zapping.conf.  (Note: interaction with nxtvepg was implemented in
  zapping too: you can check it out from CVS at zapping.sourceforge.net)
- bugfix in acq stats (control menu): when connected to acquisition daemon
  VPS/PDC info was not displayed
- added work-around for daylight saving time changes and local time zone
  handling specifically for Turkish provider TRT.
- made programme list's context menu fully configurable: pre-defined
  commands for filter and reminder control can be re-ordered now too
  (in addition to external commands); also removed the need to press the
  dialog's "Update" button after changes (which was a long standing entry
  in the TODO list)
- UNIX/X11: display Nextview logo in main window title bar (only works for
  window managers which support the WM_ICON feature, e.g. fvwm2 or Gnome)
- bugfix: columns in the main programme list were not resizable, i.e.
  the pull-down menu appeared even when clicking onto the column header
  button's right border ("single list" layout only)
- bugfix: option "-provider merged" did not work in daemon mode (acq for
  merged database did work though, if pre-configured as default in the
  .nxtvepgrc or INI file)  Thanks to Georg for reporting this bug.
- Linux: bugfix in v4l2 support (fixes problems with PVR-350 cards.)
  Thanks to Urs Schaufelberger and Hans Verkuil for the patch.
- Windows only: fixed weekday colors in datescale: did not match those
  in the "weekday color" attribute in the programme list; thanks to
  everyone who reported this bug in the forum.

See file CHANGES.RCS for a detailed log of revisions of all files.

+++ dsdrv/saa7134.c
@@ -25,7 +25,7 @@
  *  DScaler #Id: SAA7134Source.cpp,v 1.69 2003/01/27 22:04:12 laurentg Exp #
  *  DScaler #Id: SAA7134Provider.cpp,v 1.10 2002/12/24 08:22:14 atnak Exp #
  *
- *  $Id: saa7134.c,v 1.20 2004/03/22 17:35:32 tom Exp tom $
+ *  $Id: saa7134.c,v 1.21 2004/05/22 19:48:49 tom Exp tom $
  */
 
 #define DEBUG_SWITCH DEBUG_SWITCH_VBI
@@ -460,7 +460,7 @@ static void SetPageTable( uint frameIdx )
 //
 static bool CreatePageTable( uint frameIdx )
 {
-   const WORD PAGE_MASK = (DMA_PAGE_SIZE - 1);
+   const DWORD PAGE_MASK = (DMA_PAGE_SIZE - 1);
    DWORD * pPageTable;
    DWORD   pPhysical;
    BYTE  * pUser;
